M3_99: Refactor solidity and add some improvements

+++ contracts/contracts/rateX/interfaces/IDex.sol
@@ -3,6 +3,13 @@ pragma solidity ^0.8.0;
 
 interface IDex {
 
+    /// @notice Swap tokenIn for tokenOut on single dex
+    /// @param _tokenIn The contract address of the input token
+    /// @param _tokenOut The contract address of the output token
+    /// @param _amountIn The amount of tokenIn for swap
+    /// @param _amountOutMin The minimum acceptable amount of tokenOut
+    /// @param _to The recipient address of the swap
+    /// @return amountOut The amount of tokenOut received
     function swap(
         address _poolAddress,
         address _tokenIn,
@@ -10,5 +17,5 @@ interface IDex {
         uint _amountIn,
         uint _amountOutMin,
         address _to
-    ) external returns(uint amountOut);
+    ) external returns(uint256 amountOut);
 }
